+++ pyxel/editor/__init__.py
@@ -1,3 +1,4 @@
+import os
 import sys
 
 import pyxel
@@ -5,14 +6,19 @@ from pyxel.editor.app import App
 
 
 def run():
-    if len(sys.argv) == 2:
-        App(sys.argv[1])
+    arg = sys.argv[1] if len(sys.argv) >= 2 else ""
+
+    if arg.startswith("-"):
+        if arg == "-v" or arg == "--version":
+            print("Pyxel Editor {}".format(pyxel.VERSION))
+        else:
+            print("Usage: pyxeleditor [option] [pyxel_resource_file]")
+            print("Options:")
+            print(" -h, --help     This help text")
+            print(" -v, --version  Show version number and quit")
     else:
-        print("usage: pyxeleditor pyxel_resource_file")
-        print("")
-        print("Pyxel Editor {}".format(pyxel.VERSION))
-        print("Please specify an arbitrary file name to run Pyxel Editor")
-        print("e.g. pyxeleditor my_resource")
+        filename = arg or os.path.join(os.getcwd(), "my_resource")
+        App(filename)
 
 
 if __name__ == "__main__":